requirements.txt
python-telegram-bot==13.15
openai>=1.35.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0
Pillow>=10.2.0
python-dotenv>=1.0.0
pytz>=2023.3 